+++ Apps/VisionOS/VisionOS/APIViews/TrackEventsView.swift
@@ -37,11 +37,10 @@ struct TrackEventsView: View {
         VStack(alignment: .leading) {
             Markdown {
                 """
-                Using CustomerIO you can track any events. If there is an identified user, these events will be
-                attributed to them or to an annonymouse user otherwise.
+                You can track the events your users perform in your app. If you have already identified the user, these events are attributed with the identified user. Otherwise, events are attributed to an anonymous user.
                 An event must have a name.
 
-                Optionally, you can add any properties payload to the event as long
+                You can also add relevant properties to the event payload as long
                 as the payload type conforms to the Codable protocol.
                 [Learn more](https://customer.io/docs/sdk/ios/tracking/track-events/)
 
@@ -60,7 +59,7 @@ struct TrackEventsView: View {
 
             Button("Send event") {
                 if event.name.isEmpty {
-                    viewModel.errorMessage = "Please enter an event name"
+                    viewModel.errorMessage = "You must enter an event name"
                     return
                 }
                 onSuccess(event)